+++ src/lightly_train/_task_models/dinov2_ltdetr_object_detection/task_model.py
@@ -13,16 +13,18 @@ from typing import Any
 
 import torch
 from PIL.Image import Image as PILImage
+from pydantic import Field
 from torch import Tensor
 from torch.nn import Module
 from torchvision.transforms.v2 import functional as transforms_functional
 from typing_extensions import Self
 
+from lightly_train._configs.config import PydanticConfig
 from lightly_train._data import file_helpers
 from lightly_train._models import package_helpers
 from lightly_train._models.dinov2_vit.dinov2_vit_package import DINOV2_VIT_PACKAGE
 from lightly_train._task_models.dinov2_ltdetr_object_detection.dinov2_vit_wrapper import (
-    DINOv2ViTWrapper,
+    DINOv2STAs,
 )
 from lightly_train._task_models.object_detection_components import tiling_utils
 from lightly_train._task_models.object_detection_components.hybrid_encoder import (
@@ -38,6 +40,274 @@ from lightly_train._task_models.task_model import TaskModel
 from lightly_train.types import PathLike
 
 logger = logging.getLogger(__name__)
+
+
+class _HybridEncoderConfig(PydanticConfig):
+    in_channels: list[int]
+    feat_strides: list[int]
+    hidden_dim: int
+    use_encoder_idx: list[int]
+    num_encoder_layers: int
+    nhead: int
+    dim_feedforward: int
+    dropout: float
+    enc_act: str
+    expansion: float
+    depth_mult: float
+    act: str
+    upsample: bool = True
+
+
+class _HybridEncoderViTTestConfig(_HybridEncoderConfig):
+    in_channels: list[int] = [224, 224, 224]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 224
+    use_encoder_idx: list[int] = [1]
+    num_encoder_layers: int = 1
+    nhead: int = 1
+    dim_feedforward: int = 224
+    dropout: float = 0.0
+    enc_act: str = "gelu"
+    expansion: float = 1.0
+    depth_mult: float = 1.0
+    act: str = "silu"
+
+
+class _HybridEncoderViTSConfig(_HybridEncoderConfig):
+    in_channels: list[int] = [224, 224, 224]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 224
+    use_encoder_idx: list[int] = [2]
+    num_encoder_layers: int = 1
+    nhead: int = 8
+    dim_feedforward: int = 896
+    dropout: float = 0.0
+    enc_act: str = "gelu"
+    expansion: float = 1.0
+    depth_mult: float = 1.0
+    act: str = "silu"
+
+
+class _HybridEncoderViTBConfig(_HybridEncoderConfig):
+    in_channels: list[int] = [768, 768, 768]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 768
+    use_encoder_idx: list[int] = [2]
+    num_encoder_layers: int = 1
+    nhead: int = 8
+    dim_feedforward: int = 3072
+    dropout: float = 0.0
+    enc_act: str = "gelu"
+    expansion: float = 1.0
+    depth_mult: float = 1.0
+    act: str = "silu"
+
+
+class _HybridEncoderViTLConfig(_HybridEncoderConfig):
+    in_channels: list[int] = [1024, 1024, 1024]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 1024
+    use_encoder_idx: list[int] = [2]
+    num_encoder_layers: int = 1
+    nhead: int = 8
+    dim_feedforward: int = 4096
+    dropout: float = 0.0
+    enc_act: str = "gelu"
+    expansion: float = 1.0
+    depth_mult: float = 1.0
+    act: str = "silu"
+
+
+class _HybridEncoderViTGConfig(_HybridEncoderConfig):
+    in_channels: list[int] = [1536, 1536, 1536]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 1536
+    use_encoder_idx: list[int] = [2]
+    num_encoder_layers: int = 1
+    nhead: int = 8
+    dim_feedforward: int = 6144
+    dropout: float = 0.0
+    enc_act: str = "gelu"
+    expansion: float = 1.0
+    depth_mult: float = 1.0
+    act: str = "silu"
+
+
+class _RTDETRTransformerv2Config(PydanticConfig):
+    feat_channels: list[int]
+    feat_strides: list[int]
+    hidden_dim: int
+    num_levels: int
+    num_layers: int
+    num_queries: int
+    num_denoising: int
+    label_noise_ratio: float
+    box_noise_scale: float
+    eval_idx: int
+    num_points: list[int]
+    query_select_method: str
+
+
+class _RTDETRTransformerv2ViTSConfig(_RTDETRTransformerv2Config):
+    feat_channels: list[int] = [224, 224, 224]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 224
+    num_levels: int = 3
+    num_layers: int = 4
+    num_queries: int = 300
+    num_denoising: int = 100
+    label_noise_ratio: float = 0.5
+    box_noise_scale: float = 1.0
+    eval_idx: int = -1
+    num_points: list[int] = [3, 6, 3]
+    query_select_method: str = "default"
+    dim_feedforward: int = 1792
+
+
+class _RTDETRTransformerv2ViTBConfig(_RTDETRTransformerv2Config):
+    feat_channels: list[int] = [768, 768, 768]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 768
+    num_levels: int = 3
+    num_layers: int = 4
+    num_queries: int = 300
+    num_denoising: int = 100
+    label_noise_ratio: float = 0.5
+    box_noise_scale: float = 1.0
+    eval_idx: int = -1
+    num_points: list[int] = [3, 6, 3]
+    query_select_method: str = "default"
+    dim_feedforward: int = 6144
+
+
+class _RTDETRTransformerv2ViTLConfig(_RTDETRTransformerv2Config):
+    feat_channels: list[int] = [1024, 1024, 1024]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 1024
+    num_levels: int = 3
+    num_layers: int = 4
+    num_queries: int = 300
+    num_denoising: int = 100
+    label_noise_ratio: float = 0.5
+    box_noise_scale: float = 1.0
+    eval_idx: int = -1
+    num_points: list[int] = [3, 6, 3]
+    query_select_method: str = "default"
+    dim_feedforward: int = 8192
+
+
+class _RTDETRTransformerv2ViTGConfig(_RTDETRTransformerv2Config):
+    feat_channels: list[int] = [1536, 1536, 1536]
+    feat_strides: list[int] = [7, 14, 28]
+    hidden_dim: int = 1536
+    num_levels: int = 3
+    num_layers: int = 4
+    num_queries: int = 300
+    num_denoising: int = 100
+    label_noise_ratio: float = 0.5
+    box_noise_scale: float = 1.0
+    eval_idx: int = -1
+    num_points: list[int] = [3, 6, 3]
+    query_select_method: str = "default"
+    dim_feedforward: int = 12288
+
+
+class _BackboneWrapperViTSConfig(PydanticConfig):
+    interaction_indexes: list[int] = [5, 8, 11]
+    finetune: bool = True
+    conv_inplane: int = 28
+    hidden_dim: int = 224
+
+
+class _BackboneWrapperViTBConfig(PydanticConfig):
+    interaction_indexes: list[int] = [5, 8, 11]
+    finetune: bool = True
+    conv_inplane: int = 56
+    hidden_dim: int = 768
+
+
+class _BackboneWrapperViTLConfig(PydanticConfig):
+    interaction_indexes: list[int] = [11, 17, 23]
+    finetune: bool = True
+    conv_inplane: int = 56
+    hidden_dim: int = 1024
+
+
+class _BackboneWrapperViTGConfig(PydanticConfig):
+    interaction_indexes: list[int] = [19, 29, 39]
+    finetune: bool = True
+    conv_inplane: int = 64
+    hidden_dim: int = 1536
+
+
+class _RTDETRPostProcessorConfig(PydanticConfig):
+    num_top_queries: int = 300
+
+
+class _DINOv2LTDETRObjectDetectionConfig(PydanticConfig):
+    hybrid_encoder: _HybridEncoderConfig
+    rtdetr_transformer: _RTDETRTransformerv2Config
+    rtdetr_postprocessor: _RTDETRPostProcessorConfig
+
+
+class _DINOv2LTDETRObjectDetectionViTSConfig(_DINOv2LTDETRObjectDetectionConfig):
+    hybrid_encoder: _HybridEncoderViTSConfig = Field(
+        default_factory=_HybridEncoderViTSConfig
+    )
+    rtdetr_transformer: _RTDETRTransformerv2ViTSConfig = Field(
+        default_factory=_RTDETRTransformerv2ViTSConfig
+    )
+    rtdetr_postprocessor: _RTDETRPostProcessorConfig = Field(
+        default_factory=_RTDETRPostProcessorConfig
+    )
+    backbone_wrapper: _BackboneWrapperViTSConfig = Field(
+        default_factory=_BackboneWrapperViTSConfig
+    )
+
+
+class _DINOv2LTDETRObjectDetectionViTBConfig(_DINOv2LTDETRObjectDetectionConfig):
+    hybrid_encoder: _HybridEncoderViTBConfig = Field(
+        default_factory=_HybridEncoderViTBConfig
+    )
+    rtdetr_transformer: _RTDETRTransformerv2ViTBConfig = Field(
+        default_factory=_RTDETRTransformerv2ViTBConfig
+    )
+    rtdetr_postprocessor: _RTDETRPostProcessorConfig = Field(
+        default_factory=_RTDETRPostProcessorConfig
+    )
+    backbone_wrapper: _BackboneWrapperViTBConfig = Field(
+        default_factory=_BackboneWrapperViTBConfig
+    )
+
+
+class _DINOv2LTDETRObjectDetectionViTLConfig(_DINOv2LTDETRObjectDetectionConfig):
+    hybrid_encoder: _HybridEncoderViTLConfig = Field(
+        default_factory=_HybridEncoderViTLConfig
+    )
+    rtdetr_transformer: _RTDETRTransformerv2ViTLConfig = Field(
+        default_factory=_RTDETRTransformerv2ViTLConfig
+    )
+    rtdetr_postprocessor: _RTDETRPostProcessorConfig = Field(
+        default_factory=_RTDETRPostProcessorConfig
+    )
+    backbone_wrapper: _BackboneWrapperViTLConfig = Field(
+        default_factory=_BackboneWrapperViTLConfig
+    )
+
+
+class _DINOv2LTDETRObjectDetectionViTGConfig(_DINOv2LTDETRObjectDetectionConfig):
+    hybrid_encoder: _HybridEncoderViTGConfig = Field(
+        default_factory=_HybridEncoderViTGConfig
+    )
+    rtdetr_transformer: _RTDETRTransformerv2ViTGConfig = Field(
+        default_factory=_RTDETRTransformerv2ViTGConfig
+    )
+    rtdetr_postprocessor: _RTDETRPostProcessorConfig = Field(
+        default_factory=_RTDETRPostProcessorConfig
+    )
+    backbone_wrapper: _BackboneWrapperViTGConfig = Field(
+        default_factory=_BackboneWrapperViTGConfig
+    )
 
 
 class DINOv2LTDETRObjectDetection(TaskModel):
@@ -89,53 +359,52 @@ class DINOv2LTDETRObjectDetection(TaskModel):
         if load_weights and backbone_weights is not None:
             self.load_backbone_weights(dinov2, backbone_weights)
 
-        self.backbone: DINOv2ViTWrapper = DINOv2ViTWrapper(
+        # Get the configuration based on the model name.
+        config_mapping = {
+            "_vittest14": _DINOv2LTDETRObjectDetectionViTSConfig,
+            "vits14": _DINOv2LTDETRObjectDetectionViTSConfig,
+            "vitb14": _DINOv2LTDETRObjectDetectionViTBConfig,
+            "vitl14": _DINOv2LTDETRObjectDetectionViTLConfig,
+            "vitg14": _DINOv2LTDETRObjectDetectionViTGConfig,
+        }
+        config_name = parsed_name["backbone_name"].replace("-notpretrained", "")
+        config_name = config_name.replace("-noreg", "")
+        config_cls = config_mapping[config_name]
+        config = config_cls()
+
+        # TODO(Guarin, 02/26): Improve how mask tokens are handled for fine-tuning.
+        dinov2.mask_token.requires_grad = False  # type: ignore
+
+        self.backbone: DINOv2STAs = DINOv2STAs(
             model=dinov2,
-            keep_indices=[5, 8, 11],
+            # Disable STA for DINOv2 as it doesn't work well with patch size 14.
+            use_sta=False,
+            **config.backbone_wrapper.model_dump(),
         )
-        # TODO(Lionel, 07/25): Improve how mask tokens are handled for fine-tuning.
-        # Should we drop them from the model? We disable grads here for DDP to work
-        # without find_unused_parameters=True.
-        self.backbone.backbone.mask_token.requires_grad = False
 
         self.encoder: HybridEncoder = HybridEncoder(  # type: ignore[no-untyped-call]
-            in_channels=[384, 384, 384],
-            feat_strides=[14, 14, 14],
-            hidden_dim=384,
-            use_encoder_idx=[2],
-            num_encoder_layers=1,
-            nhead=8,
-            dim_feedforward=2048,
-            dropout=0.0,
-            enc_act="gelu",
-            expansion=1.0,
-            depth_mult=1,
-            act="silu",
-            upsample=False,
+            **config.hybrid_encoder.model_dump()
         )
 
+        decoder_config = config.rtdetr_transformer.model_dump()
+        decoder_config.update({"num_classes": len(self.classes)})
         self.decoder: RTDETRTransformerv2 = RTDETRTransformerv2(  # type: ignore[no-untyped-call]
-            num_classes=len(self.classes),
-            feat_channels=[384, 384, 384],
-            feat_strides=[14, 14, 14],
-            hidden_dim=256,
-            num_levels=3,
-            num_layers=6,
-            num_queries=300,
-            num_denoising=100,
-            label_noise_ratio=0.5,
-            box_noise_scale=1.0,
-            eval_idx=-1,
-            num_points=[4, 4, 4],
-            query_select_method="default",
-            # TODO Lionel (09/25): Remove when anchors are not in checkpoints anymore.
+            **decoder_config,
             eval_spatial_size=self.image_size,  # From global config, otherwise anchors are not generated.
         )
 
+        postprocessor_config = config.rtdetr_postprocessor.model_dump()
+        postprocessor_config.update({"num_classes": len(self.classes)})
         self.postprocessor: RTDETRPostProcessor = RTDETRPostProcessor(
-            num_classes=len(self.classes),
-            num_top_queries=300,
+            **postprocessor_config
         )
+
+    @classmethod
+    def list_model_names(cls) -> list[str]:
+        return [
+            f"{name}-{cls.model_suffix}"
+            for name in DINOV2_VIT_PACKAGE.list_model_names()
+        ]
 
     @classmethod
     def is_supported_model(cls, model: str) -> bool:
@@ -181,12 +450,27 @@ class DINOv2LTDETRObjectDetection(TaskModel):
             "backbone_name": backbone_name,
         }
 
-    @classmethod
-    def list_model_names(cls) -> list[str]:
-        return [
-            f"{name}-{cls.model_suffix}"
-            for name in DINOV2_VIT_PACKAGE.list_model_names()
-        ]
+    def load_train_state_dict(
+        self, state_dict: dict[str, Any], strict: bool = True, assign: bool = False
+    ) -> Any:
+        """Load the state dict from a training checkpoint.
+
+        Loads the EMA weights if available, otherwise falls back to the model weights.
+        """
+        has_ema_weights = any(k.startswith("ema_model.model.") for k in state_dict)
+        has_model_weights = any(k.startswith("model.") for k in state_dict)
+        new_state_dict = {}
+        if has_ema_weights:
+            for name, param in state_dict.items():
+                if name.startswith("ema_model.model."):
+                    name = name[len("ema_model.model.") :]
+                    new_state_dict[name] = param
+        elif has_model_weights:
+            for name, param in state_dict.items():
+                if name.startswith("model."):
+                    name = name[len("model.") :]
+                    new_state_dict[name] = param
+        return self.load_state_dict(new_state_dict, strict=strict, assign=assign)
 
     def load_backbone_weights(self, backbone: Module, path: PathLike) -> None:
         """
@@ -216,27 +500,13 @@ class DINOv2LTDETRObjectDetection(TaskModel):
         else:
             logger.info(f"Backbone weights loaded from '{path}'")
 
-    def load_train_state_dict(
-        self, state_dict: dict[str, Any], strict: bool = True, assign: bool = False
-    ) -> Any:
-        """Load the state dict from a training checkpoint.
-
-        Loads the EMA weights if available, otherwise falls back to the model weights.
-        """
-        has_ema_weights = any(k.startswith("ema_model.model.") for k in state_dict)
-        has_model_weights = any(k.startswith("model.") for k in state_dict)
-        new_state_dict = {}
-        if has_ema_weights:
-            for name, param in state_dict.items():
-                if name.startswith("ema_model.model."):
-                    name = name[len("ema_model.model.") :]
-                    new_state_dict[name] = param
-        elif has_model_weights:
-            for name, param in state_dict.items():
-                if name.startswith("model."):
-                    name = name[len("model.") :]
-                    new_state_dict[name] = param
-        return self.load_state_dict(new_state_dict, strict=strict, assign=assign)
+    def deploy(self) -> Self:
+        self.eval()
+        self.postprocessor.deploy()  # type: ignore[no-untyped-call]
+        for m in self.modules():
+            if hasattr(m, "convert_to_deploy"):
+                m.convert_to_deploy()  # type: ignore[operator]
+        return self
 
     @torch.no_grad()
     def predict(
@@ -414,20 +684,6 @@ class DINOv2LTDETRObjectDetection(TaskModel):
             "scores": scores,
         }
 
-    def deploy(self) -> Self:
-        self.eval()
-        self.postprocessor.deploy()  # type: ignore[no-untyped-call]
-        for m in self.modules():
-            if hasattr(m, "convert_to_deploy"):
-                m.convert_to_deploy()  # type: ignore[operator]
-        return self
-
-    def _forward_train(self, x: Tensor, targets):  # type: ignore[no-untyped-def]
-        x = self.backbone(x)
-        x = self.encoder(x)
-        x = self.decoder(feats=x, targets=targets)
-        return x
-
     def forward(
         self, x: Tensor, orig_target_size: Tensor | None = None
     ) -> tuple[Tensor, Tensor, Tensor]:
@@ -456,6 +712,12 @@ class DINOv2LTDETRObjectDetection(TaskModel):
         labels, boxes, scores = result
         labels = self.internal_class_to_class[labels]
         return (labels, boxes, scores)
+
+    def _forward_train(self, x: Tensor, targets):  # type: ignore[no-untyped-def]
+        x = self.backbone(x)
+        x = self.encoder(x)
+        x = self.decoder(feats=x, targets=targets)
+        return x
 
 
 class DINOv2LTDETRDSPObjectDetection(DINOv2LTDETRObjectDetection):
@@ -499,48 +761,38 @@ class DINOv2LTDETRDSPObjectDetection(DINOv2LTDETRObjectDetection):
             model_name=parsed_name["backbone_name"],
             model_args=backbone_args,
         )
-        self.backbone: DINOv2ViTWrapper = DINOv2ViTWrapper(
+
+        # Get the configuration based on the model name.
+        config_mapping = {
+            "vits14": _DINOv2LTDETRObjectDetectionViTSConfig,
+            "vitb14": _DINOv2LTDETRObjectDetectionViTBConfig,
+            "vitl14": _DINOv2LTDETRObjectDetectionViTLConfig,
+            "vitg14": _DINOv2LTDETRObjectDetectionViTGConfig,
+        }
+        config_name = parsed_name["backbone_name"]
+        config_cls = config_mapping[config_name]
+        config = config_cls()
+
+        self.backbone: DINOv2STAs = DINOv2STAs(
             model=dinov2,
-            keep_indices=[5, 8, 11],
+            # Disable STA for DINOv2 as it doesn't work well with patch size 14.
+            use_sta=False,
+            **config.backbone_wrapper.model_dump(),
         )
 
         self.encoder: HybridEncoder = HybridEncoder(  # type: ignore[no-untyped-call]
-            in_channels=[384, 384, 384],
-            feat_strides=[14, 14, 14],
-            hidden_dim=384,
-            use_encoder_idx=[2],
-            num_encoder_layers=1,
-            nhead=8,
-            dim_feedforward=2048,
-            dropout=0.0,
-            enc_act="gelu",
-            expansion=1.0,
-            depth_mult=1,
-            act="silu",
-            upsample=False,
+            **config.hybrid_encoder.model_dump()
         )
 
+        decoder_config = config.rtdetr_transformer.model_dump()
+        decoder_config.update({"cross_attn_method": "discrete"})
+        decoder_config.update({"num_classes": len(self.classes)})
         self.decoder: RTDETRTransformerv2 = RTDETRTransformerv2(  # type: ignore[no-untyped-call]
-            feat_channels=[384, 384, 384],
-            feat_strides=[14, 14, 14],
-            hidden_dim=256,
-            num_levels=3,
-            cross_attn_method="discrete",
-            num_layers=6,
-            num_queries=300,
-            num_denoising=100,
-            label_noise_ratio=0.5,
-            box_noise_scale=1.0,
-            eval_idx=-1,
-            num_points=[4, 4, 4],
-            query_select_method="default",
-            # TODO Lionel (09/25): Remove when anchors are not in checkpoints anymore.
-            eval_spatial_size=(
-                644,
-                644,
-            ),  # From global config, otherwise anchors are not generated.
+            **decoder_config,
+            eval_spatial_size=self.image_size,  # From global config, otherwise anchors are not generated.
         )
 
+        postprocessor_config = config.rtdetr_postprocessor.model_dump()
         self.postprocessor: RTDETRPostProcessor = RTDETRPostProcessor(
-            num_top_queries=300,
+            **postprocessor_config
         )