+++ CHANGELOG.md
@@ -9,6 +9,13 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- LTDETR support for DINOv3 ViT-B/L and DINOv2 ViT-L/B/G models. Pretrained weights are
+  not yet available for these models.
+- Add support for fine-tuning DINOv2 models for instance segmentation with the
+  `train_instance_segmentation` command. See the
+  [instance segmentation documentation](https://docs.lightly.ai/train/stable/instance_segmentation.html#model)
+  for more information.
+
 ### Changed
 
 ### Deprecated
@@ -16,6 +23,8 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Removed
 
 ### Fixed
+
+- Filter invalid bounding boxes in instance segmentation
 
 ### Security